refactor: Rename AQ.StateMachineEntities to AQ.StateMachine.Entities for consistency
