+++ src/StateMachine/Entities/AQ.StateMachine.Entities.csproj
