+++ src/StateMachine/Services/IStateMachineRequirementContextService.cs
@@ -1,0 +1,6 @@
+namespace AQ.StateMachine.Services;
+
+public interface IStateMachineRequirementContextService
+{
+
+}
