+++ src/StateMachine/Services/StateMachineEffectExecutionService.cs
@@ -1,4 +1,4 @@
-﻿using AQ.StateMachineEntities;
+using AQ.StateMachine.Entities;
 using System.Diagnostics;
 using System.Reflection;
 
@@ -359,3 +359,4 @@ public class StateMachineEffectExecutionService : IStateMachineEffectExecutionSe
         }
     }
 }
+
