+++ src/StateMachine/Services/StateMachineServiceCollectionExtensions.cs
@@ -18,3 +18,4 @@ public static class StateMachineServiceCollectionExtensions
         return services;
     }
 }
+
