+++ tests/StateMachine/Services/AQ.StateMachine.Services.Tests.csproj
@@ -26,7 +26,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\..\src\StateMachine\Services\AQ.StateMachine.Services.csproj" />
-    <ProjectReference Include="..\..\..\src\StateMachine\Entities\AQ.StateMachineEntities.csproj" />
+    <ProjectReference Include="..\..\..\src\StateMachine\Entities\AQ.StateMachine.Entities.csproj" />
     <ProjectReference Include="..\..\..\src\Utilities\Results\AQ.Utilities.Results.csproj" />
     <ProjectReference Include="..\..\..\src\Abstractions\AQ.Abstractions.csproj" />
   </ItemGroup>
